modified:   Makefile.am 		added sample directory
	modified:   README
		mention sample programs

	modified:   expgram/Discount.hpp
		tweaded debug messages

	modified:   progs/ngram_index.cpp
		added quantize option

	new file:   sample/Makefile.am
	new file:   sample/ngram.cpp
	new file:   sample/ngram_counts.cpp
		addeds ample program

	modified:   utils/pfor_vector.hpp
	modified:   utils/sample_pfor_vector.cpp
		larger buffer size with pfor

+++ expgram/Discount.hpp
@@ -79,13 +79,13 @@ namespace expgram
     double lower_order_weight(const count_type total, const count_type observed, const count_type min2, const count_type min3) const
     { 
       if (modified)
-       return (discount1 >= 0.0 && discount2 >= 0.0 && discount3plus > 0.0
-	       ? (discount1 * (observed - min2) + discount2 * (min2 - min3) + discount3plus * min3) / total // modified-KN smoothing
-	       : observed / (total + observed)); // witten-bell smoothing
-     else
-       return (discount1 >= 0.0
-	       ? (discount1 * observed) / total    // KN smoothing
-	       : double(observed) / (total + observed)); // witten-bell smoothing
+	return (discount1 >= 0.0 && discount2 >= 0.0 && discount3plus > 0.0
+		? (discount1 * (observed - min2) + discount2 * (min2 - min3) + discount3plus * min3) / total // modified-KN smoothing
+		: double(observed) / (total + observed)); // witten-bell smoothing
+      else
+	return (discount1 >= 0.0
+		? (discount1 * observed) / total    // KN smoothing
+		: double(observed) / (total + observed)); // witten-bell smoothing
     }
     
     double discount(const count_type count, const count_type total, const count_type observed) const
